[WIP] fix: Properly adjust padding on dock
Font Awesome adds a random amount of margin to all the icons on the
right. This causes a lot of oddities when looking at the dock's padding,
hence ruining the whole experience.

In order to fix this, adjust 'margin-right' and 'margin-left', if required,
to all the elements on the dock. As well, set '-webkit-tap-highlight-color'
to 'transparent' because this hack adds very weird-looking boxes when
tapping on the text.